Offset changes back to 32

+++ Source/core/IPCMessage.h
@@ -285,14 +285,14 @@ namespace Core {
             inline void Set (const uint16_t length, const uint8_t buffer[]) {
                 _buffer.Copy(0, length, buffer);
             }
-            inline uint16_t Serialize(uint8_t buffer[], const uint16_t length, const uint16_t offset) const
+            inline uint16_t Serialize(uint8_t buffer[], const uint16_t length, const uint32_t offset) const
             {
                 uint16_t size = ((_buffer.Size() - offset) > length ? length : (_buffer.Size() - offset));
                 ::memcpy(buffer, &(_buffer[offset]), size);
 
                 return (size);
             }
-            uint16_t Deserialize(const uint8_t buffer[], const uint16_t length, const uint16_t offset)
+            uint16_t Deserialize(const uint8_t buffer[], const uint16_t length, const uint32_t offset)
             {
                 ASSERT (offset == _buffer.Size());
